Completed assignment

+++ server/app.py
@@ -35,19 +35,25 @@ def bakeries():
 
 @app.route('/bakeries/<int:id>', methods=['GET', 'PATCH'])
 def bakery_by_id(id):
-    bakery = Bakery.query.filter_by(id=id).first()
-    if request.method == 'GET':
 
-        bakery_serialized = bakery.to_dict()
+    bakery = Bakery.query.filter_by(id=id).first()
+
+    if request.method == 'GET':
+        bakery_dict = bakery.to_dict()
 
         response = make_response(
-            jsonify(bakery_serialized),
+            jsonify(bakery_dict),
             200
         )
+
         return response
+    
     elif request.method == 'PATCH':
+        bakery = Bakery.query.filter_by(id=id).first()
+
         for attr in request.form:
             setattr(bakery, attr, request.form.get(attr))
+
         db.session.add(bakery)
         db.session.commit()
 
@@ -60,26 +66,53 @@ def bakery_by_id(id):
 
         return response
 
+@app.route('/baked_goods/by_price')
+def baked_goods_by_price():
+    baked_goods_by_price = BakedGood.query.order_by(BakedGood.price).all()
+    baked_goods_by_price_serialized = [
+        bg.to_dict() for bg in baked_goods_by_price
+    ]
+    
+    response = make_response(
+        baked_goods_by_price_serialized,
+        200
+    )
+    return response
+
+@app.route('/baked_goods/most_expensive')
+def most_expensive_baked_good():
+    most_expensive = BakedGood.query.order_by(BakedGood.price.desc()).limit(1).first()
+    most_expensive_serialized = most_expensive.to_dict()
+
+    response = make_response(
+        most_expensive_serialized,
+        200
+    )
+    return response
+
 
 @app.route('/baked_goods', methods=['GET', 'POST'])
-def baked_goods():
+def post_baked_goods():
     if request.method == 'GET':
         baked_goods = []
         for baked_good in BakedGood.query.all():
-            baked_goods_dict = baked_good.to_dict()
-            baked_goods.append(baked_goods_dict)
+            baked_good_dict = baked_good.to_dict()
+            baked_goods.append(baked_good_dict)
+
         response = make_response(
             jsonify(baked_goods),
             200
         )
-        return response
 
+        return response
+    
     elif request.method == 'POST':
         new_baked_good = BakedGood(
-            name=request.form.get('name'),
-            price=request.form.get('price'),
-            bakery_id=request.form.get('bakery_id')
+            name=request.form.get("name"),
+            price=request.form.get("price"),
+            bakery_id=request.form.get("bakery_id"),
         )
+
         db.session.add(new_baked_good)
         db.session.commit()
 
@@ -91,49 +124,40 @@ def baked_goods():
         )
 
         return response
+    
 
-@app.route('/baked_goods/<int:id>', methods=['DELETE'])
-def delete_baked_goods(id):
-    baked_good = BakedGood.query.filter(BakedGood.id == id).first()
-    db.session.delete(baked_good)
-    db.session.commit()
+@app.route('/baked_goods/<int:id>', methods=['GET', 'DELETE'])
+def baked_goods_by_id(id):
+    baked_good = BakedGood.query.filter_by(id=id).first()
 
-    response_body = {
-        "delete_successful": True,
-        "message": "Review deleted."
-    }
-    response = make_response(
-        response_body,
-        200
-    )
-
-    return response
-
-@app.route('/baked_goods/by_price')
-def baked_goods_by_price():
-    baked_goods_by_price = BakedGood.query.order_by(BakedGood.price).all()
-    baked_goods_by_price_serialized = [
-        bg.to_dict() for bg in baked_goods_by_price
-    ]
-
-    response = make_response(
-        baked_goods_by_price_serialized,
-        200
-    )
-    return response
+    if request.method == 'GET':
+        baked_good_dict = baked_good.to_dict()
 
 
-@app.route('/baked_goods/most_expensive')
-def most_expensive_baked_good():
-    most_expensive = BakedGood.query.order_by(
-        BakedGood.price.desc()).limit(1).first()
-    most_expensive_serialized = most_expensive.to_dict()
+        response = make_response(
+            jsonify(baked_good_dict),
+            200
+        )
 
-    response = make_response(
-        most_expensive_serialized,
-        200
-    )
-    return response
+        return response
+    
+
+    elif request.method == 'DELETE':
+        db.session.delete(baked_good)
+        db.session.commit()
+
+        response_body = {
+            "delete_successful": True,
+            "message": "Baked good deleted."    
+        }
+
+        response = make_response(
+            jsonify(response_body),
+            200
+        )
+
+        return response
+
 
 
 if __name__ == '__main__':
